--- tests/Input/Week/RenderTest.php
@@ -1,397 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace PHPForge\Html\Tests\Input\Week;
-
-use PHPForge\Html\Input\Week;
-use PHPForge\Support\Assert;
-use PHPUnit\Framework\TestCase;
-
-/**
- * @psalm-suppress PropertyNotSetInConstructor
- */
-final class RenderTest extends TestCase
-{
-    public function testAriaDescribedBy(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week" aria-describedby="MyWidget">
-            HTML,
-            Week::widget()->ariaDescribedBy('MyWidget')->id('week-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testAriaLabel(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week" aria-label="MyWidget">
-            HTML,
-            Week::widget()->ariaLabel('MyWidget')->id('week-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testAttributes(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input class="class" id="week-6582f2d099e8b" type="week">
-            HTML,
-            Week::widget()->attributes(['class' => 'class'])->id('week-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testAutofocus(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week" autofocus>
-            HTML,
-            Week::widget()->autofocus()->id('week-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testClass(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input class="class" id="week-6582f2d099e8b" type="week">
-            HTML,
-            Week::widget()->class('class')->id('week-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testDataAttributes(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week" data-test="data-value">
-            HTML,
-            Week::widget()->dataAttributes(['test' => 'data-value'])->id('week-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testDisabled(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week" disabled>
-            HTML,
-            Week::widget()->disabled()->id('week-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testForm(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week" form="form">
-            HTML,
-            Week::widget()->form('form')->id('week-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testGenerateAriaDescribeBy(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week" aria-describedby="week-6582f2d099e8b-help">
-            HTML,
-            Week::widget()->ariaDescribedBy()->id('week-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testGenerateAriaDescribeByWithFalse(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week">
-            HTML,
-            Week::widget()->ariaDescribedBy(false)->id('week-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testHidden(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week" hidden>
-            HTML,
-            Week::widget()->hidden()->id('week-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testId(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="id" type="week">
-            HTML,
-            Week::widget()->id('id')->render()
-        );
-    }
-
-    public function testLang(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week" lang="en">
-            HTML,
-            Week::widget()->id('week-6582f2d099e8b')->lang('en')->render()
-        );
-    }
-
-    public function testName(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" name="name" type="week">
-            HTML,
-            Week::widget()->id('week-6582f2d099e8b')->name('name')->render()
-        );
-    }
-
-    public function testPrefix(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            prefix
-            <input id="week-6582f2d099e8b" type="week">
-            HTML,
-            Week::widget()->id('week-6582f2d099e8b')->prefix('prefix')->render()
-        );
-    }
-
-    public function testPrefixContainer(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div>
-            prefix
-            </div>
-            <input id="week-6582f2d099e8b" type="week">
-            HTML,
-            Week::widget()->id('week-6582f2d099e8b')->prefix('prefix')->prefixContainer(true)->render()
-        );
-    }
-
-    public function testPrefixContainerAttributes(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div class="class">
-            prefix
-            </div>
-            <input id="week-6582f2d099e8b" type="week">
-            HTML,
-            Week::widget()
-                ->id('week-6582f2d099e8b')
-                ->prefix('prefix')
-                ->prefixContainer(true)
-                ->prefixContainerAttributes(['class' => 'class'])
-                ->render()
-        );
-    }
-
-    public function testPrefixContainerClass(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div class="class">
-            prefix
-            </div>
-            <input id="week-6582f2d099e8b" type="week">
-            HTML,
-            Week::widget()
-                ->id('week-6582f2d099e8b')
-                ->prefix('prefix')
-                ->prefixContainer(true)
-                ->prefixContainerClass('class')
-                ->render()
-        );
-    }
-
-    public function testPrefixContainerTag(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <article>
-            prefix
-            </article>
-            <input id="week-6582f2d099e8b" type="week">
-            HTML,
-            Week::widget()
-                ->id('week-6582f2d099e8b')
-                ->prefix('prefix')
-                ->prefixContainer(true)
-                ->prefixContainerTag('article')
-                ->render()
-        );
-    }
-
-    public function testReadonly(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week" readonly>
-            HTML,
-            Week::widget()->id('week-6582f2d099e8b')->readonly()->render()
-        );
-    }
-
-    public function testRender(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week">
-            HTML,
-            Week::widget()->id('week-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testStyle(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week" style="style">
-            HTML,
-            Week::widget()->id('week-6582f2d099e8b')->style('style')->render()
-        );
-    }
-
-    public function testSuffix(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week">
-            suffix
-            HTML,
-            Week::widget()->id('week-6582f2d099e8b')->suffix('suffix')->render()
-        );
-    }
-
-    public function testSuffixContainer(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week">
-            <div>
-            suffix
-            </div>
-            HTML,
-            Week::widget()->id('week-6582f2d099e8b')->suffix('suffix')->suffixContainer(true)->render()
-        );
-    }
-
-    public function testSuffixContainerAttributes(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week">
-            <div class="class">
-            suffix
-            </div>
-            HTML,
-            Week::widget()
-                ->id('week-6582f2d099e8b')
-                ->suffix('suffix')
-                ->suffixContainer(true)
-                ->suffixContainerAttributes(['class' => 'class'])
-                ->render()
-        );
-    }
-
-    public function testSuffixContainerClass(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week">
-            <div class="class">
-            suffix
-            </div>
-            HTML,
-            Week::widget()
-                ->id('week-6582f2d099e8b')
-                ->suffix('suffix')
-                ->suffixContainer(true)
-                ->suffixContainerClass('class')
-                ->render()
-        );
-    }
-
-    public function testSuffixContainerTag(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week">
-            <article>
-            suffix
-            </article>
-            HTML,
-            Week::widget()
-                ->id('week-6582f2d099e8b')
-                ->suffix('suffix')
-                ->suffixContainer(true)
-                ->suffixContainerTag('article')
-                ->render()
-        );
-    }
-
-    public function testTabIndex(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week" tabindex="1">
-            HTML,
-            Week::widget()->id('week-6582f2d099e8b')->tabIndex(1)->render()
-        );
-    }
-
-    public function testTemplate(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week">
-            suffix
-            HTML,
-            Week::widget()
-                ->id('week-6582f2d099e8b')
-                ->prefix('prefix')
-                ->suffix('suffix')
-                ->template('{tag}\n{suffix}')
-                ->render()
-        );
-    }
-
-    public function testTitle(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week" title="title">
-            HTML,
-            Week::widget()->id('week-6582f2d099e8b')->title('title')->render()
-        );
-    }
-
-    public function testValue(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week" value="value">
-            HTML,
-            Week::widget()->id('week-6582f2d099e8b')->value('value')->render()
-        );
-    }
-
-    public function testValueWithEmpty(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week">
-            HTML,
-            Week::widget()->id('week-6582f2d099e8b')->value(null)->render()
-        );
-    }
-}
